requirements.txt
python-telegram-bot
dotenv
pandas
openpyxl
python-docx